Idea for improving forward compatibility
I've run into some dependencies that use the RetryPolicy constructor
from 0.6 as well as getRetryPolicy. This has been renamed in 0.7, but
because RetryPolicy is a type alias, we could retain the same
constructor name and newtype accessor to make upgrades easier. I have
unfortunately found that because of the polymorphic m on RetryPolicyM
that the user will have to add RankNTypes if they don't have it already,
but no actual code will need to change.