Another take on #17 with RetryStatus
Predicting another future request of getting more stats in the handler
and monadic action, we make a retry policy take a RetryStatus which
contains the retry number and cumulative delay so far. We cannot include
retries remaining or delay remaining due to the fact that RetryPolicy is
a monad. I note that this changes lots of Int types to RetryStatus so
the effort to prevent breakage is for naught, as the only people who
would be able to upgrade without changes are ones who used "const" with
no type annotations of Int to be found.