Update docs, add previous delay, lens RetryStatus
This should be ready for review now. We decided to not export the
constructor of RetryStatus and instead export the field and generic
lenses. The lenses don't incur any dependencies on lens and should be
compatible with the common lens tools out there.

Because of the type change required by RetryStatus, we already lost the
battle on breaking changes so I opted to not add specialized versions of
retrying, recovering, and recoverAll and just make the change requested
in #17.