Add special cases for async exceptions.
We generally don't want retry catching async exceptions. The current
best way of handling this at the moment seems to be the
enclosed-exceptions package, but we didn't want to introduce the
additional dependencies in this release. Now, recoverAll will explicitly
not catch any AsyncException or SomeAsyncException if available. If the
user throws a custom async exception then it will still get caught, so
documentation to that effect has been added.