Resolve test warnings
This resolves #82. The code in that MR worked but it produced warnings
in the current default lts for the project. I used the qualified import
trick to resolve the warning. Then the partial matches started failing
in newer GHC versions so I resolved that too. This now builds without
warnings on the project LTS, the newest LTS and nightly.